requirements.txt
openai-whisper
imageio-ffmpeg